manage menu, restore user

+++ content/moduls.php
@@ -3,13 +3,14 @@
     $id_role = isset($_SESSION['ID_ROLE'])? $_SESSION['ID_ROLE'] : '';
 
     $rowStudent = mysqli_fetch_assoc(mysqli_query($config, "SELECT * FROM students WHERE id = '$id_user'"));
-    $id_majors = $rowStudent['id_majors'];
+    $id_majors = isset($rowStudent['id_majors'])? $rowStudent['id_majors'] : '';
     
-    if($id_role == 2){
+    if($id_role == 8){
         $where = "WHERE moduls.id_majors='$id_majors'";
-    }else if($id_role == 1){
+    }else if($id_role == 6){
         $where = "WHERE moduls.id_instructors='$id_user'";
     }
+
     $queryModuls = mysqli_query($config, "SELECT majors.name as majors_name, instructors.name as instructors_name, moduls.* 
     FROM moduls 
     LEFT JOIN majors ON majors.id = moduls.id_majors
@@ -19,13 +20,13 @@
     $rowModuls = mysqli_fetch_all($queryModuls, MYSQLI_ASSOC);
     // print_r($rowModuls);die;
 ?>
-
+ 
 <div class="row">
   <div class="col-12">
       <div class="card">
           <div class="card-body">
               <h5 class="card-title mt-3">Data Modul</h5>
-              <?php if(isset($_SESSION['ID_ROLE']) && $_SESSION['ID_ROLE'] == 1):?>
+              <?php if(canAddModul($id_role)):?>
                 <div class="mb-3" align="right">
                     <a href="?page=tambah-moduls" class="btn btn-primary mb-3" >Add Modul</a>
                 </div>
@@ -38,7 +39,9 @@
                               <th>Title</th>
                               <th>Instructors</th>
                               <th>Majors</th>
-                              <th>Action</th>
+                              <?php if($id_role==6):?>
+                                <th>Action</th>
+                                <?php endif ?>
                           </tr>
                       </thead>
                       <tbody>
@@ -53,10 +56,12 @@
                               </td>
                               <td><?= $row['instructors_name'] ?></td>
                               <td><?= $row['majors_name'] ?></td>
-                              <td>
-                                  <a href="?page=tambah-modul&edit=<?php echo $row['id']?>" class = "btn btn-primary" name="edit">Edit</a>
-                                  <a onclick="return confirm('Are you sure wanna delete this data?')" href="?page=tambah-moduls&delete=<?php echo $row['id']?>" class = "btn btn-danger" name="delete">Delete</a>
-                              </td>
+                              <?php if($id_role==6):?>
+                                <td>
+                                    <a href="?page=tambah-modul&edit=<?php echo $row['id']?>" class = "btn btn-primary" name="edit">Edit</a>
+                                    <a onclick="return confirm('Are you sure wanna delete this data?')" href="?page=tambah-moduls&delete=<?php echo $row['id']?>" class = "btn btn-danger" name="delete">Delete</a>
+                                </td>
+                              <?php endif ?>
                           </tr>
                           <?php endforeach?>
                       </tbody>
